requirements.txt
Django==1.11
djangorestframework==3.6.2
olefile==0.44
pytz==2017.2
Pillow==4.0.0